feat(l2): S2 硬化修复 + 恢复 Stage 3 budget 接线 + IMP-1a rollout temperature
- S2 硬化（服务器实测修复）：refresh 相位 IS 权重 log_ratio_clip（REFRESH_LOG_RATIO_MAX=3.0）、
  base 路径 BASE_LOG_RATIO_CLIP=5.0、staleness_drop_base（base 池跳过陈旧度截断）、
  bf16 显存修复（s_old/s_cur/delta 转 bf16）、P1.5 真实 pad_token_id、load_cache 补加载
  教师对（L2 刷新相位需要）、compute_disagreement 硬 gate、KL 锚点 T 截断。
- 恢复 Stage 3（Budget-Aware Selective Rollout）：assign_budgets/enforce_budget/
  compute_rollout_metrics/group_by_budget 纯函数 + RefreshSelector.select_with_budget +
  _value 补 reward 项 + DynamicRatioController token 感知 + pipeline budget 接线。
  修复工作区残留的部分回滚导致 test_l2_budget 收集失败（约束4：不删除 correctness test）。
- IMP-1a：l2.rollout.temperature 默认 0.7（pipeline 不写死），run_refresh_phase 完整透传
  到注入生成器/默认生成器，summary 记录 temperature，rollout/temperature 指标落盘，
  =1.0 可复现旧行为。373 tests green（此前 341）。

+++ main/fullstack_opd_v2/cache_store.py
@@ -132,7 +132,9 @@ def write_cache_disk(cache, prefix: str, responses: torch.Tensor | None = None,
     N, T, K = cache.ids_sorted.shape
 
     ids_np = cache.ids_sorted.cpu().numpy().astype(np.int32)
-    delta_np = cache.delta_k_sorted.cpu().numpy().astype(np.float32)
+    delta_np = cache.delta_k_sorted.float().cpu().numpy().astype(np.float32)
+    # 注：delta_k_sorted 源自 teacher bf16 前向 → bf16 张量，numpy 不支持 bf16，
+    # 必须先 .float() 再 numpy（否则 `Got unsupported ScalarType BFloat16`）。
     # 逐 chunk 直写（避免整文件 numpy 副本驻留 RAM）
     _write_memmap(f"{prefix}.ids_sorted.dat", ids_np, chunk)
     _write_memmap(f"{prefix}.delta_k_sorted.dat", delta_np, chunk)
